Change from float to double

+++ src/grdio.h
@@ -5,12 +5,12 @@
  *
  */
 int grdread (const char *filename, size_t *p_nx, size_t *p_ny, 
-		      float **pp_x, float **pp_y, float **pp_z);
+		      double **pp_x, double **pp_y, double **pp_z);
 
 /**
  *
  */
 int grdwrite (const char *filename, const size_t nx, size_t ny, 
-				  const float *p_x, const float *p_y, const float *p_z);
+				  const double *p_x, const double *p_y, const double *p_z);
 
 #endif
